# server/docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  Login:
    properties:
      password:
        description: The password to login
        type: string
        x-go-name: Password
      username:
        description: The username to login
        type: string
        x-go-name: Username
    required:
    - username
    - password
    type: object
    x-go-package: github.com/Davidmnj91/myrents/pkg/login
  Register:
    properties:
      birth_date:
        description: The BirthDate of the user
        type: string
        x-go-name: BirthDate
      email:
        description: The Email of the user
        type: string
        x-go-name: Email
      id_number:
        description: The IDNumber of the user
        type: string
        x-go-name: IDNumber
      name:
        description: The Name of the user
        type: string
        x-go-name: Name
      password:
        description: The Password of the user
        type: string
        x-go-name: Password
      phone:
        description: The Phone of the user
        type: string
        x-go-name: Phone
      surname:
        description: The Surname of the user
        type: string
        x-go-name: Surname
      user_name:
        description: The name of the user in the app
        type: string
        x-go-name: Username
    required:
    - user_name
    - password
    - name
    - surname
    - id_number
    - email
    - phone
    - birth_date
    type: object
    x-go-package: github.com/Davidmnj91/myrents/pkg/user_register
host: localhost:5000
info:
  description: Documentation of our MyRents API.
  title: classification MyRents.
  version: 1.0.0
paths:
  /account:
    delete:
      operationId: remove-account
      responses:
        "200":
          description: Successful delete
        "500":
          description: Internal server error
      summary: Deletes an existing account on the system.
      tags:
      - user
  /login:
    post:
      operationId: login-user
      parameters:
      - description: Body to log in a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Login'
      responses:
        "200":
          description: Successful login, bearer token in "authorization" header
        "500":
          description: Internal server error
      summary: Logs in a user account into the system.
      tags:
      - auth
  /logout:
    delete:
      operationId: logout
      responses:
        "200":
          description: Successful logout
        "500":
          description: Internal server error
      summary: Logs out a user account from the system.
      tags:
      - auth
  /register:
    post:
      operationId: register-user
      parameters:
      - description: Body to register a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Register'
      responses:
        "200":
          description: Successful registration
        "500":
          description: Internal server error
      summary: Creates a new user in the system.
      tags:
      - user
produces:
- application/json
schemes:
- http
swagger: "2.0"